requirements.txt
timm==0.5.4 
einops==0.4.1
fvcore
easydict
matplotlib
numpy==1.21.0
yacs
scikit-image==0.19.3
pillow
wandb
torch==2.4.1
scikit-learn>=1.0